pyproject.toml
[tool.poetry]
name = "Vehicle VIN lookup"
version = "0.1.0"
description = "Example of FastAPI"
authors = ["Erik Rivera (erik@rivera.pro)"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.95.0"
uvicorn = "0.22.0"
pydantic = {extras = ["dotenv"], version = "^1.10.7"}
requests = "^2.22.0"
sqlalchemy = "^2.0.13"
aiosqlite = "^0.19.0"
pyarrow = "^12.0.0"
pandas = "^2.0.1"
pyjwt = "^2.7.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = {version = "^19.10b0", allow-prereleases = true}
autopep8 = "^2.0.0"
ipdb = "^0.13.0"
pylint = "^2.15.8"
httpx = "^0.24.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.0.0"
requests-mock = "^1.10.0"
pytest-freezegun = "^0.4.2"

[tool.poetry.group.aws]
optional = true

[tool.poetry.group.aws.dependencies]
mangum = "^0.17.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 120
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
  \.toml
  |\.sh
  |\.git
  |\.ini
  |Dockerfile
  |Jenkinfile
)/
'''

[tool.pytest.ini_options]
asyncio_mode = 'auto'